Part 1 Complete!
The first part of my Space SHMUP project is complete! The second part will be done in due time. I also want to notify that my repo is SpaceSHMUPPart1, but that the final build will be called SpaceSHMUP. I've followed everything to a tee, and that included the repo name, so hopefully that's alright.